requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
gunicorn
requests
python-dotenv
psycopg2-binary
Flask-Caching
